move concatenate function to utils.py
